remove_bg/requirements.txt
albumentations==0.4.5
altair==4.2.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
imageio==2.22.0
imgaug==0.2.6
importlib-metadata==4.12.0
Jinja2==3.1.2
jsonschema==4.16.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.1
networkx==2.8.6
numpy==1.23.3
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf==3.20.1
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
rich==12.5.1
scikit-image==0.19.3
scipy==1.9.1
semver==2.13.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.0
streamlit==1.13.0
tifffile==2022.8.12
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
zipp==3.8.1
